migration fix

+++ migrations/003_booking_tables.js
@@ -15,7 +15,7 @@ const up = async (knex) => {
 			t.date('startDate').notNullable();
 			t.date('endDate').notNullable();
 			t.timestamp('createdAt').defaultTo(knex.fn.now());
-			t.tinyint('statusId').unsigned().defaultTo(6)
+			t.integer('statusId').unsigned().defaultTo(6)
 				.references('id').inTable('BookingStatuses').onDelete('SET NULL');
 			t.text('additionalInfo');
 		});
